# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KOSA Review Analysis API",
    "version": "1.0.0",
    "description": "리뷰 기반 실시간 감정 분석 및 요약 서비스 API",
    "contact": {
      "name": "KOSA Team",
      "email": "support@kosa.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.kosa.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 토큰을 사용한 인증"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "에러 메시지"
          },
          "code": {
            "type": "string",
            "description": "에러 코드"
          },
          "details": {
            "type": "object",
            "description": "에러 상세 정보"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/info/changelog": {
      "get": {
        "summary": "API 변경 이력 조회",
        "description": "API의 버전별 변경 이력을 조회합니다.",
        "tags": [
          "API Info"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string"
            },
            "description": "시작 버전 (예: 1.0.0)"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string"
            },
            "description": "종료 버전 (예: 1.1.0)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "조회할 최대 항목 수"
          }
        ],
        "responses": {
          "200": {
            "description": "변경 이력 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "changelog": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "version": {
                            "type": "string"
                          },
                          "date": {
                            "type": "string",
                            "format": "date"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "major",
                              "minor",
                              "patch"
                            ]
                          },
                          "changes": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "enum": [
                                    "added",
                                    "changed",
                                    "deprecated",
                                    "removed",
                                    "fixed",
                                    "security"
                                  ]
                                },
                                "description": {
                                  "type": "string"
                                },
                                "endpoints": {
                                  "type": "array",
                                  "items": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "전체 변경 이력 수"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/info/version/{version}": {
      "get": {
        "summary": "특정 버전 정보 조회",
        "description": "특정 API 버전의 상세 정보를 조회합니다.",
        "tags": [
          "API Info"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "조회할 버전 (예: v1)"
          }
        ],
        "responses": {
          "200": {
            "description": "버전 정보 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "version": {
                      "type": "string"
                    },
                    "releaseDate": {
                      "type": "string",
                      "format": "date"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "stable",
                        "beta",
                        "deprecated"
                      ]
                    },
                    "description": {
                      "type": "string"
                    },
                    "endpoints": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "deprecationDate": {
                      "type": "string",
                      "format": "date",
                      "nullable": true
                    },
                    "supportEndDate": {
                      "type": "string",
                      "format": "date",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "버전을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}